Move all latestDependencies files to main, delete unityIAP files
